test(lsp): merge the common and harness test modules into one harness
tests/harness/mod.rs (session plumbing) and tests/common/mod.rs
(Published normalization + CLI comparison) were two single-file modules
each carrying #![allow(dead_code)] because every test binary compiled
only a subset. Fold common into harness, and fold
configuration_refresh.rs's private spawn_session (distinct only in
taking client capabilities and initializationOptions) in as
spawn_session_with_options so the duplicate spawner is gone. Pure code
motion; no test behavior changes.
